post requirements +fix
toggle to require op file on imageboard
toggle to require op subject on imageboard (textboard already has by default )

if no subject is given, check if neither image or comment is given. if true, error. if subject is not required, check if both image and subjectcomment is empty. you need at least one of the three to post a thread.

also changes error message for some length settings to be dynamic and not hardcoded